+++ scripts/update-go-version.sh
@@ -5,7 +5,7 @@
 latest=$(curl -s 'https://go.dev/dl/?mode=json' | jq -r '.[].stable = true | .[].version' | head -n 1)
 version="${latest#go}"
 if [[ "${latest}" =~ ^go ]]; then
-  sed -i -re "s/^FROM golang:.*/FROM golang:${version}-alpine3.18 AS build/" Dockerfile
+  sed -i -re "s/^FROM golang:.*/FROM golang:${version}-alpine3.19 AS build/" Dockerfile
   for f in .github/workflows/*.yml; do
     sed -i -re "s/GOVERSION: .*/GOVERSION: '>=${version}'/" $f
   done